Remove unused preview support files to streamline the project structure. This includes the deletion of BuilderExamples.swift, BuilderPreviews.swift, and PreviewModels.swift, which were not utilized in the current implementation, enhancing overall code organization and maintainability.
