--- Sources/DynamicList/PreviewSupport/BuilderPreviews.swift
@@ -1,251 +1,0 @@
-//
-//  Copyright © 2025 Jesús Alfredo Hernández Alarcón. All rights reserved.
-//
-
-import Combine
-import SwiftUI
-
-// MARK: - Preview Data Models
-
-struct PreviewUser: Identifiable, Hashable {
-    let id = UUID()
-    let name: String
-    let email: String
-    let avatar: String
-
-    static let sampleUsers = [
-        PreviewUser(name: "Ana García", email: "ana@example.com", avatar: "👩‍💼"),
-        PreviewUser(name: "Carlos López", email: "carlos@example.com", avatar: "👨‍💻"),
-        PreviewUser(name: "María Rodríguez", email: "maria@example.com", avatar: "👩‍🎨"),
-        PreviewUser(name: "Juan Pérez", email: "juan@example.com", avatar: "👨‍🔬"),
-    ]
-}
-
-struct PreviewProduct: Identifiable, Hashable {
-    let id = UUID()
-    let name: String
-    let price: Double
-    let category: String
-
-    static let sampleProducts = [
-        PreviewProduct(name: "iPhone 15", price: 999.99, category: "Electronics"),
-        PreviewProduct(name: "MacBook Pro", price: 1999.99, category: "Computers"),
-        PreviewProduct(name: "AirPods Pro", price: 249.99, category: "Audio"),
-        PreviewProduct(name: "iPad Air", price: 599.99, category: "Tablets"),
-    ]
-}
-
-// MARK: - Builder Previews
-
-#Preview("Simple Builder") {
-    DynamicListBuilder<PreviewUser>()
-        .items(PreviewUser.sampleUsers)
-        .title("Usuarios")
-        .rowContent { user in
-            HStack {
-                Text(user.avatar)
-                    .font(.title2)
-                VStack(alignment: .leading) {
-                    Text(user.name)
-                        .font(.headline)
-                    Text(user.email)
-                        .font(.caption)
-                        .foregroundColor(.secondary)
-                }
-                Spacer()
-            }
-            .padding(.vertical, 4)
-        }
-        .detailContent { user in
-            VStack(spacing: 20) {
-                Text(user.avatar)
-                    .font(.system(size: 80))
-
-                Text(user.name)
-                    .font(.title)
-                    .fontWeight(.bold)
-
-                Text(user.email)
-                    .font(.body)
-                    .foregroundColor(.secondary)
-
-                Spacer()
-            }
-            .padding()
-            .navigationTitle(DynamicListPresenter.profile)
-        }
-        .build()
-}
-
-#Preview("Reactive Builder") {
-    let productsPublisher: AnyPublisher<[PreviewProduct], Error> = Just(PreviewProduct.sampleProducts)
-        .delay(for: .seconds(1.0), scheduler: DispatchQueue.main)
-        .setFailureType(to: Error.self)
-        .eraseToAnyPublisher()
-
-    return DynamicListBuilder<PreviewProduct>()
-        .publisher(productsPublisher)
-        .title("Productos")
-        .rowContent { product in
-            HStack {
-                VStack(alignment: .leading) {
-                    Text(product.name)
-                        .font(.headline)
-                    Text(product.category)
-                        .font(.caption)
-                        .foregroundColor(.secondary)
-                }
-                Spacer()
-                Text("$\(product.price, specifier: "%.2f")")
-                    .font(.subheadline)
-                    .fontWeight(.semibold)
-                    .foregroundColor(.blue)
-            }
-            .padding(.vertical, 4)
-        }
-        .detailContent { product in
-            VStack(spacing: 20) {
-                Text("📦")
-                    .font(.system(size: 80))
-
-                Text(product.name)
-                    .font(.title)
-                    .fontWeight(.bold)
-
-                Text(product.category)
-                    .font(.headline)
-                    .foregroundColor(.secondary)
-
-                Text("$\(product.price, specifier: "%.2f")")
-                    .font(.title2)
-                    .fontWeight(.bold)
-                    .foregroundColor(.blue)
-
-                Spacer()
-            }
-            .padding()
-            .navigationTitle(DynamicListPresenter.detail)
-        }
-        .build()
-}
-
-#Preview("Simulated Loading") {
-    DynamicListBuilder<PreviewUser>()
-        .simulatedPublisher(PreviewUser.sampleUsers, delay: 2.0)
-        .title("Cargando Usuarios")
-        .rowContent { user in
-            HStack {
-                Text(user.avatar)
-                Text(user.name)
-                    .font(.body)
-                Spacer()
-            }
-        }
-        .detailContent { user in
-            Text("Detalle de \(user.name)")
-                .navigationTitle(DynamicListPresenter.userDetail)
-        }
-        .build()
-}
-
-#Preview("Custom Error View") {
-    let failingPublisher: AnyPublisher<[PreviewUser], Error> = Fail(error: NSError(domain: "Example", code: 404, userInfo: [
-        NSLocalizedDescriptionKey: "No se pudieron cargar los usuarios",
-    ]))
-    .delay(for: .seconds(1.0), scheduler: DispatchQueue.main)
-    .eraseToAnyPublisher()
-
-    return DynamicListBuilder<PreviewUser>()
-        .publisher(failingPublisher)
-        .title("Usuarios")
-        .rowContent { user in
-            Text(user.name)
-        }
-        .detailContent { user in
-            Text("Detalle de \(user.name)")
-        }
-        .errorContent { error in
-            VStack(spacing: 16) {
-                Text("😞")
-                    .font(.system(size: 60))
-
-                Text("¡Ups! Algo salió mal")
-                    .font(.title2)
-                    .fontWeight(.bold)
-
-                Text(error.localizedDescription)
-                    .font(.body)
-                    .foregroundColor(.secondary)
-                    .multilineTextAlignment(.center)
-
-                Button(DynamicListPresenter.retry) {
-                    // This would trigger a refresh
-                }
-                .buttonStyle(.borderedProminent)
-            }
-            .padding()
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-        }
-        .build()
-}
-
-#Preview("Default Views") {
-    DynamicListBuilder<PreviewUser>()
-        .items(PreviewUser.sampleUsers)
-        .title("Vistas por Defecto")
-        .build()
-}
-
-// MARK: - Factory Method Previews
-
-#Preview("Simple Factory") {
-    DynamicListBuilder.simple(
-        items: PreviewUser.sampleUsers,
-        rowContent: { user in
-            HStack {
-                Text(user.avatar)
-                Text(user.name)
-                Spacer()
-            }
-        },
-        detailContent: { user in
-            Text("Detalle de \(user.name)")
-                .navigationTitle("Usuario")
-        },
-    )
-}
-
-#Preview("Reactive Factory") {
-    let publisher: AnyPublisher<[PreviewProduct], Error> = Just(PreviewProduct.sampleProducts)
-        .delay(for: .seconds(1.0), scheduler: DispatchQueue.main)
-        .setFailureType(to: Error.self)
-        .eraseToAnyPublisher()
-
-    return DynamicListBuilder.reactive(
-        publisher: publisher,
-        rowContent: { product in
-            HStack {
-                Text(product.name)
-                Spacer()
-                Text("$\(product.price, specifier: "%.2f")")
-            }
-        },
-        detailContent: { product in
-            Text("Detalle de \(product.name)")
-                .navigationTitle("Producto")
-        },
-    )
-}
-
-#Preview("Simulated Factory") {
-    DynamicListBuilder.simulated(
-        items: PreviewUser.sampleUsers,
-        delay: 2.0,
-        rowContent: { user in
-            Text(user.name)
-        },
-        detailContent: { user in
-            Text("Detalle de \(user.name)")
-        },
-    )
-}
